fix(electron): unpack native modules from ASAR
