+++ package.json
@@ -144,6 +144,7 @@
     "@electron-forge/maker-dmg": "7.8.0",
     "@electron-forge/maker-squirrel": "7.8.0",
     "@electron-forge/maker-zip": "7.8.0",
+    "@electron-forge/plugin-auto-unpack-natives": "7.8.0",
     "@electron/rebuild": "3.7.2",
     "@reforged/maker-appimage": "5.1.1",
     "@types/lodash": "4.17.24",
